requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1
sqlmodel==0.0.16
bcrypt>=4.0.1
passlib[bcrypt]
PyJWT==2.8.0
aiosmtplib==4.0.1
